# app/launchpad/launchpad_api/models/user_request.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from .base_model import Model
from .. import util


class UserRequest(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, name=None, emailid=None, role=None):  # noqa: E501
        """UserRequest - a model defined in OpenAPI

        :param name: The name of this UserRequest.  # noqa: E501
        :type name: str
        :param emailid: The emailid of this UserRequest.  # noqa: E501
        :type emailid: str
        :param role: The role of this UserRequest.  # noqa: E501
        :type role: int
        """
        self.openapi_types = {
            'name': str,
            'emailid': str,
            'role': int
        }

        self.attribute_map = {
            'name': 'name',
            'emailid': 'emailid',
            'role': 'role'
        }

        self._name = name
        self._emailid = emailid
        self._role = role

    @classmethod
    def from_dict(cls, dikt) -> 'UserRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UserRequest of this UserRequest.  # noqa: E501
        :rtype: UserRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this UserRequest.


        :return: The name of this UserRequest.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this UserRequest.


        :param name: The name of this UserRequest.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def emailid(self) -> str:
        """Gets the emailid of this UserRequest.


        :return: The emailid of this UserRequest.
        :rtype: str
        """
        return self._emailid

    @emailid.setter
    def emailid(self, emailid: str):
        """Sets the emailid of this UserRequest.


        :param emailid: The emailid of this UserRequest.
        :type emailid: str
        """
        if emailid is None:
            raise ValueError("Invalid value for `emailid`, must not be `None`")  # noqa: E501

        self._emailid = emailid

    @property
    def role(self) -> int:
        """Gets the role of this UserRequest.


        :return: The role of this UserRequest.
        :rtype: int
        """
        return self._role

    @role.setter
    def role(self, role: int):
        """Sets the role of this UserRequest.


        :param role: The role of this UserRequest.
        :type role: int
        """
        if role is None:
            raise ValueError("Invalid value for `role`, must not be `None`")  # noqa: E501

        self._role = role