# app/launchpad/launchpad_api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Dashboard API
  version: 1.0.0
servers:
- url: http://localhost:8080/api
security:
- cookieAuth: []
paths:
...
  /user:
    delete:
      operationId: user_delete
      parameters:
      - explode: true
        in: query
        name: user_id
        required: true
        schema:
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: User deleted successfully
      summary: Delete a user
      tags:
      - user
      x-openapi-router-controller: app.launchpad.launchpad_api.controllers.user_controller
    get:
      operationId: user_get
      parameters:
      - explode: true
        in: query
        name: id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: List of users fetched successfully
      summary: Get list of users
      tags:
      - user
      x-openapi-router-controller: app.launchpad.launchpad_api.controllers.user_controller
    post:
      operationId: user_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: User created successfully
      summary: Create a new user
      tags:
      - user
      x-openapi-router-controller: app.launchpad.launchpad_api.controllers.user_controller
    put:
      operationId: user_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: User updated successfully
      summary: Update an existing user
      tags:
      - user
      x-openapi-router-controller: app.launchpad.launchpad_api.controllers.user_controller
...
  /user/{user_id}:
    get:
      operationId: user_by_id_get
      parameters:
      - explode: false
        in: path
        name: user_id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: User fetched successfully
      summary: Get user by ID
      tags:
      - user
      x-openapi-router-controller: app.launchpad.launchpad_api.controllers.user_controller
    put:
      operationId: user_by_id_put
      parameters:
      - explode: false
        in: path
        name: user_id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: User updated successfully
      summary: Update user by ID
      tags:
      - user
      x-openapi-router-controller: app.launchpad.launchpad_api.controllers.user_controller
    delete:
      operationId: user_by_id_delete
      parameters:
      - explode: false
        in: path
        name: user_id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: User deleted successfully
      summary: Delete user by ID
      tags:
      - user
      x-openapi-router-controller: app.launchpad.launchpad_api.controllers.user_controller
...
components:
  schemas:
...
    UserRequest:
      example:
        role: 0
        name: John Doe
        emailid: john.doe@example.com
      properties:
        name:
          example: John Doe
          title: name
          type: string
        emailid:
          example: john.doe@example.com
          format: email
          title: emailid
          type: string
        role:
          title: role
          type: integer
      required:
      - emailid
      - name
      - role
      title: UserRequest
      type: object
...
  securitySchemes:
    cookieAuth:
      in: cookie
      name: session_id
      type: apiKey
      x-apikeyInfoFunc: app.launchpad.launchpad_api.controllers.security_controller.info_from_cookieAuth